Isolated requests from backend
This commit isolates the requests methods inside a dedicated file for
backend.
