+++ src/server/remote/handlers.hpp
@@ -1,6 +1,8 @@
 #ifndef CAPIO_SERVER_REMOTE_HANDLERS_HPP
 #define CAPIO_SERVER_REMOTE_HANDLERS_HPP
 
+#include "handlers/requests.hpp"
+
 #include "handlers/read.hpp"
 #include "handlers/stat.hpp"
